Pull Datasource out of interfaces.py for improved TypeVar behavior
