+++ tests/datasource/fluent/test_spark_filesystem_datasource.py
@@ -5,7 +5,7 @@ import logging
 import pathlib
 import re
 from dataclasses import dataclass
-from typing import List
+from typing import TYPE_CHECKING, List
 
 import pytest
 
@@ -15,7 +15,6 @@ from great_expectations.datasource.fluent.data_asset.data_connector import (
     FilesystemDataConnector,
 )
 from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition,
     TestConnectionError,
 )
 from great_expectations.datasource.fluent.spark_file_path_datasource import (
@@ -24,6 +23,9 @@ from great_expectations.datasource.fluent.spark_file_path_datasource import (
 from great_expectations.datasource.fluent.spark_filesystem_datasource import (
     SparkFilesystemDatasource,
 )
+
+if TYPE_CHECKING:
+    from great_expectations.datasource.fluent.datasource import SortersDefinition
 
 logger = logging.getLogger(__name__)
